[Scrapper] Начать слушать топики Outbox

+++ scrapper/src/main/java/ru/overcode/scrapper/service/link/LinkDbService.java
@@ -3,6 +3,7 @@ package ru.overcode.scrapper.service.link;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
 import ru.overcode.scrapper.model.link.Link;
 import ru.overcode.scrapper.repository.link.LinkRepository;
 
@@ -25,11 +26,17 @@ public class LinkDbService {
 
     @Transactional
     public void deleteAllByIdInBatch(Collection<Long> linkIds) {
+        if (CollectionUtils.isEmpty(linkIds)) {
+            return;
+        }
         linkRepository.deleteAllByIdInBatch(linkIds);
     }
 
     @Transactional(readOnly = true)
     public Map<Long, Link> findAllById(Collection<Long> linkIds) {
+        if (CollectionUtils.isEmpty(linkIds)) {
+            return Map.of();
+        }
         return linkRepository.findAllById(linkIds).stream()
                 .collect(Collectors.toMap(
                         Link::getId,
